Created data access module with functions that interact with the state directly. Modified all other modules to interact with the data access module.
