+++ scripts/DiamondSizes.js
@@ -1,4 +1,4 @@
-import { getOrderBuilder, getSizes, setSize } from "./database.js"
+import { getOrderBuilder, getSizes, setSize } from "./dataAccess.js"
 import { renderAllHTML } from "./main.js"
 
 const sizes = getSizes()
