+++ scripts/database.js
@@ -1,5 +1,3 @@
-import { renderAllHTML } from "./main.js"
-
 /*
 
     This module contains all of the data, or state, for the
@@ -7,7 +5,7 @@ import { renderAllHTML } from "./main.js"
     modules to get copies of the state.
 
 */
-const database = {
+export const database = {
     styles: [
         { id: 1, style: "Classic", price: 500 },
         { id: 2, style: "Modern", price: 710 },
@@ -56,48 +54,4 @@ const database = {
     ],
     orderBuilder:
         {}
-}
-
-export const getMetals = () => {
-    return database.metals.map(metal => ({ ...metal }))
-}
-export const getSizes = () => {
-    return database.sizes.map(size => ({ ...size }))
-}
-export const getStyles = () => {
-    return database.styles.map(style => ({ ...style }))
-}
-export const getOrders = () => {
-    return database.customOrders.map(order => ({ ...order }))
-}
-export const getOrderBuilder = () => {
-    return database.orderBuilder
-}
-export const getTypes = () => {
-    return database.types.map(type => ({ ...type }))
-}
-
-
-// Order builder for setting temporary state.
-export const setMetal = (id) => {
-    database.orderBuilder.metalId = id
-}
-export const setSize = (id) => {
-    database.orderBuilder.sizeId = id
-}
-export const setStyle = (id) => {
-    database.orderBuilder.styleId = id
-}
-export const setType = (id) => {
-    database.orderBuilder.typeId = id
-}
-
-export const addCustomOrder = () => {
-    const newOrder = { ...database.orderBuilder }
-    const lastIndex = database.customOrders.length - 1
-    newOrder.id = database.customOrders[lastIndex].id + 1
-    newOrder.timestamp = Date.now()
-    database.customOrders.push(newOrder)
-    database.orderBuilder = {}
-    document.dispatchEvent(new CustomEvent("stateChanged"))
 }
